+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ###Unreleased/pending:
 * Removed all forced/hard focus due to undesired experience for non-disabled users.
 * Removed WASD keyboard press events to traversing the slides from v2.7.0 as they were conflicting with screen reader keyboard events.
+* Removed all tab indexes
+* Updated sr-only elements
+* Display the "OPEN" word on the splash screen rather than a play icon
+* Changed semantic of the word "play" to "open"
 
 ---
 ###2.7.0 (08-25-2015)